requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
Pillow==11.1.0
google-genai==1.14.0
pydantic==2.10.4
